app/src/main/java/event_horizon/service/impl/EventServiceImpl.java
package event_horizon.service.impl;

import event_horizon.exception.EventNotFoundException;
import event_horizon.model.Event;
import event_horizon.repository.EventRepository;
import event_horizon.service.EventService;
import lombok.NonNull;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class EventServiceImpl implements EventService {
    private final EventRepository eventRepository;

    @Autowired
    public EventServiceImpl(EventRepository eventRepository) {
        this.eventRepository = eventRepository;
    }

    @Override
    public Event getEventById(@NonNull Long id) {
        return eventRepository.findById(id).orElseThrow(() -> new EventNotFoundException(id));
    }
}
